Split off the horrible TransduceEvent code into a separate mod
